fix: use empty event title and add placeholder
